32 bit variance

+++ src/boot/ksdk1.1.0/pedometer.c
@@ -33,7 +33,7 @@ int32_t compute_mean(int16_t data[], int16_t N){
 }
 
 
-int64_t compute_variance(int16_t data[], int32_t mean, int8_t N){
+int32_t compute_variance(int16_t data[], int32_t mean, int8_t N){
 	int64_t sum = 0;
 	SEGGER_RTT_printf(0, "\nMean = %ld \t Mean^2 = %ld\n", mean, mean*mean);
 	for (int i=0; i<N; i++){
